videoAPI/requirements.txt
django==4.0
djangorestframework==3.13.1
markdown==3.3.6
django-filter==21.1
apscheduler==3.8.1
PyYAML==6.0
requests==2.26.0
django-filter==1.1.0